Removed hosts table, spaces now use users table and dev seed data reflects this, also done a bit of folder restructure as i went along
