+++ app.py
@@ -91,7 +91,7 @@ def create_space():
 @app.route("/register", methods=["GET", "POST"])
 def register():
     if request.method == "GET":
-        return render_template("Users/register.html")
+        return render_template("users/register.html")
 
     connection = get_flask_database_connection(app)
     user_repo = UserRepository(connection)
@@ -106,7 +106,7 @@ def register():
     # Check if any field is empty
     if not all([first_name, last_name, email, contact_number, password]):
         errors = "All fields are required. Please fill in all fields."
-        return render_template("Users/register.html", errors=errors)
+        return render_template("users/register.html", errors=errors)
 
     # Hash the password before passing to DB
     hashed_password = bcrypt.generate_password_hash(password).decode("utf-8")
@@ -124,7 +124,7 @@ def register():
 @app.route("/login", methods=["GET", "POST"])
 def login():
     if request.method == "GET":
-        return render_template("Users/login.html")
+        return render_template("users/login.html")
 
     connection = get_flask_database_connection(app)
     user_repo = UserRepository(connection)
@@ -142,7 +142,7 @@ def login():
         login_user(user_session)
         return redirect("/spaces")
 
-    return render_template("Users/login.html", errors="Incorrect email or password.")
+    return render_template("users/login.html", errors="Incorrect email or password.")
 
 
 @app.route("/logout")
