+++ seeds/dev_seed_all.sql
@@ -1,16 +1,17 @@
 DROP TABLE IF EXISTS spaces;
 DROP SEQUENCE IF EXISTS spaces_id_seq;
-DROP TABLE IF EXISTS hosts;
-DROP SEQUENCE IF EXISTS hosts_id_seq;
+DROP TABLE IF EXISTS users;
+DROP SEQUENCE IF EXISTS users_id_seq;
 
-CREATE SEQUENCE IF NOT EXISTS hosts_id_seq;
-CREATE TABLE hosts (
+CREATE SEQUENCE IF NOT EXISTS users_id_seq;
+CREATE TABLE users (
     id SERIAL PRIMARY KEY,
     first_name VARCHAR(255),
     last_name VARCHAR(255),
+    email VARCHAR(255),
     contact_number VARCHAR(20),
-    email VARCHAR(255)
-    );
+    password VARCHAR(255)
+);
 
 CREATE SEQUENCE IF NOT EXISTS spaces_id_seq;
 CREATE TABLE spaces (
@@ -28,12 +29,12 @@ CREATE TABLE spaces (
     host_id INTEGER,
     CONSTRAINT fk_host FOREIGN KEY (host_id)
     REFERENCES hosts(id) ON DELETE CASCADE
-    );
+);
 
-INSERT INTO hosts (first_name, last_name, contact_number, email) VALUES ('Bilbo', 'Baggins', '0789 123 8765', 'bilbob59@hobbitmail.org');
-INSERT INTO hosts (first_name, last_name, contact_number, email) VALUES ('Mario', 'Mario', '0800 100 400', 'mario@warioland.com');
-INSERT INTO hosts (first_name, last_name, contact_number, email) VALUES ('Luke', 'Skywalker', '0208 783 1234', 'notvadersson@jedi.net');
-INSERT INTO hosts (first_name, last_name, contact_number, email) VALUES ('Rubeus', 'Hagrid', '0207 934 0232', 'keeperofkeys@hogwarts.edu');
+INSERT INTO users (first_name, last_name, contact_number, email, password) VALUES ('Bilbo', 'Baggins', '0789 123 8765', 'bilbob59@hobbitmail.org', '$2b$12$eD4USQ4Mgd7RUc88c9UqOe6oddbqnVlXDqgguOrLtVESjpbRqUn2.');
+INSERT INTO users (first_name, last_name, contact_number, email, password) VALUES ('Mario', 'Mario', '0800 100 400', 'mario@warioland.com', '$2b$12$eD4USQ4Mgd7RUc88c9UqOe6oddbqnVlXDqgguOrLtVESjpbRqUn2.');
+INSERT INTO users (first_name, last_name, contact_number, email, password) VALUES ('Luke', 'Skywalker', '0208 783 1234', 'notvadersson@jedi.net', '$2b$12$eD4USQ4Mgd7RUc88c9UqOe6oddbqnVlXDqgguOrLtVESjpbRqUn2.');
+INSERT INTO users (first_name, last_name, contact_number, email, password) VALUES ('Rubeus', 'Hagrid', '0207 934 0232', 'keeperofkeys@hogwarts.edu', '$2b$12$eD4USQ4Mgd7RUc88c9UqOe6oddbqnVlXDqgguOrLtVESjpbRqUn2.');
 
 INSERT INTO spaces (property_name, location, beds, property_type, price_per_night, description, image_url, rating, availability, booked_dates, host_id)
 	VALUES('Bag End', 'Hobbiton', 9, 'House', 289.50, 'Step into the heart of the Shire and experience the timeless charm of Bag End. This cozy,
