+++ templates/spaces/single_space.html
@@ -20,7 +20,7 @@
 					<strong class="star">⭐️</strong>
                 {% endfor %}
 			</span>
-            <p class="space-description">{{space.description}}</p>
+            <p class="space-description">{{ space.description }}</p>
         </div>
     </div>
     <div class="dates-available">
