--- tests/test_app.py
@@ -1,3 +1,0 @@
-from playwright.sync_api import Page, expect
-
-# Tests for your routes go here
