+++ tests/test_integration.py
@@ -5,10 +5,8 @@ from lib.spaces.space_repo import *
 """
 Test that list_of_spaces.html template renders with all the information in the database
 """
-
-
 def test_get_spaces(db_connection, page, test_web_address):
-    db_connection.seed("seeds/spaces_library.sql")
+    db_connection.seed("seeds/tests/spaces_test_seed.sql")
 
     # Check header text
     page.goto(f"http://{test_web_address}/spaces")
@@ -33,54 +31,52 @@ def test_get_spaces(db_connection, page, test_web_address):
 
 # ==================== /spaces/new route ====================
 
-"""
-Tests that form to create a new space can be submitted 
-"""
+# """
+# Tests that form to create a new space can be submitted 
+# """
+# def test_create_new_space_form_exists(page, test_web_address):
+#     # Go to the /space/new page
+#     page.goto(f"http://{test_web_address}/spaces/new")
+#     print(page.content())
 
+#     # Check header text
+#     header_h1 = page.locator("h1")
+#     expect(header_h1).to_have_text("Create New Space")
 
-def test_create_new_space_form_exists(page, test_web_address):
-    # Go to the /space/new page
-    page.goto(f"http://{test_web_address}/spaces/new")
+#     # Ensure that the form exists
+#     form = page.locator("form")
+#     expect(form).to_be_visible()
 
-    # Check header text
-    header_h1 = page.locator("h1")
-    expect(header_h1).to_have_text("Create New Space")
+# def test_create_new_space(db_connection, page, test_web_address):
+#     db_connection.seed("seeds/tests/spaces_test_seed.sql")
 
-    # Ensure that the form exists
-    form = page.locator("form")
-    expect(form).to_be_visible()
+#     # Go to the /space/new page
+#     page.goto(f"http://{test_web_address}/spaces/new")
 
+#     # Fill out the form
+#     page.fill("input[name=property_name]", "london flat")
+#     page.fill("input[name=price_per_night]", "30")
+#     page.fill("input[name=beds]", "2")
+#     page.fill("input[name=location]", "London")
+#     page.fill(
+#         "input[name=image_url]",
+#         "https://images.prismic.io/tembo/Zwff6YF3NbkBXMYy_semidetachedhouseuk.jpg",
+#     )
+#     page.locator("[name=property_type]").select_option("House")
+#     page.fill("input[name=rating]", "5")
+#     page.fill("#description", "cosy london apartment")
+#     page.check("#availability")
+#     page.fill("input[name=host_name]", "Bill")
+#     page.click("text=Submit")
 
-def test_create_new_space(db_connection, page, test_web_address):
-    db_connection.seed("seeds/spaces_library.sql")
+#     #     # Check that we're redirected back to the /spaces page
+#     #     expect(page).to_have_url(f"http://{test_web_address}/spaces")
 
-    # Go to the /space/new page
-    page.goto(f"http://{test_web_address}/spaces/new")
-
-    # Fill out the form
-    page.fill("input[name=property_name]", "london flat")
-    page.fill("input[name=price_per_night]", "30")
-    page.fill("input[name=beds]", "2")
-    page.fill("input[name=location]", "London")
-    page.fill(
-        "input[name=image_url]",
-        "https://images.prismic.io/tembo/Zwff6YF3NbkBXMYy_semidetachedhouseuk.jpg",
-    )
-    page.locator("[name=property_type]").select_option("House")
-    page.fill("input[name=rating]", "5")
-    page.fill("#description", "cosy london apartment")
-    page.check("#availability")
-    page.fill("input[name=host_name]", "Bill")
-    page.click("text=Submit")
-
-    #     # Check that we're redirected back to the /spaces page
-    #     expect(page).to_have_url(f"http://{test_web_address}/spaces")
-
-    # Now check that the new space is added to the /spaces route
-    space_details = page.locator(".space-item .space-details").last
-    expect(space_details).to_contain_text("london flat")
-    expect(space_details).to_contain_text("£30.00 per night")
-    expect(space_details).to_contain_text("2 beds")
-    expect(space_details).to_contain_text("House in London")
-    stars = space_details.locator(".star")
-    assert stars.count() == 5
+#     # Now check that the new space is added to the /spaces route
+#     space_details = page.locator(".space-item .space-details").last
+#     expect(space_details).to_contain_text("london flat")
+#     expect(space_details).to_contain_text("£30.00 per night")
+#     expect(space_details).to_contain_text("2 beds")
+#     expect(space_details).to_contain_text("House in London")
+#     stars = space_details.locator(".star")
+#     assert stars.count() == 5
